pyproject.toml
[tool.poetry]
name = "orgdigestor"
version = "0.1.0"
description = ""
authors = ["Ernesto Mancilla <emancillao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
markdown = "^3.6"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"


[tool.poetry.group.dev.dependencies]
pywatchman = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
